Revert "Merge branch 'development' into development-Yuming"
This reverts commit c43ab21af9c5713bbc856dbf47805d5ab311bd83, reversing
changes made to 53d47f2fbcc3fea4148f8d6f9cef684768c5e69c.
