--- Groundbreakers/Assets/Enemies.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a35fb8252e80bb94ba1e772803f337b2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
