--- Groundbreakers/Assets/Enemies/Fire Bat.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c32cbb9f84ea5334eba84ab943dd62e4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
