--- Groundbreakers/Assets/Enemies/Fire_Bat_Behavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 900e0aa16df566542bcb79b2deb0db3d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
