--- Groundbreakers/Assets/Scenes/Tiles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 296b93783c4d10d4d9828e3a4c0c9556
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
