--- Groundbreakers/Assets/Scripts/TerrainSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d647538d09194da4a8e0df18799e6a43
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
