--- Groundbreakers/Assets/Scripts/TerrainSystemEditor.cs
@@ -1,15 +1,0 @@
-﻿using UnityEditor;
-
-namespace Assets.Scripts
-{
-    [CustomEditor(typeof(TerrainSystem))]
-    public class TerrainSystemEditor : Editor
-    {
-        public override void OnInspectorGUI()
-        {
-            this.DrawDefaultInspector();
-
-            EditorGUILayout.HelpBox("This is a help box", MessageType.Info);
-        }
-    }
-}
