--- Groundbreakers/Assets/Scripts/TerrainSystemEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ae8fa6025faefc4fa1ec36cb9e0da7a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
